+++ src/fallback-features/completion-provider.ts
@@ -1,10 +1,11 @@
-import { CancellationToken, TextDocument, Position, Hover } from 'vscode';
-import * as fs from 'fs';
 import * as vscode from 'vscode';
-import { isPositionInString, FORTRAN_KEYWORDS } from '../util/helper';
-import { getDeclaredFunctions } from './functions';
-import { EXTENSION_ID } from '../util/tools';
+import { Position, TextDocument } from 'vscode';
+
 import { Logger } from '../services/logging';
+import { FORTRAN_KEYWORDS, isPositionInString } from '../util/helper';
+import { EXTENSION_ID } from '../util/tools';
+
+import { getDeclaredFunctions } from './functions';
 import intrinsics from './intrinsics.json';
 
 class CaseCoverter {
