+++ src/fallback-features/hover-provider.ts
@@ -1,5 +1,7 @@
 import { CancellationToken, TextDocument, Position, Hover } from 'vscode';
+
 import { Logger } from '../services/logging';
+
 import intrinsics from './intrinsics.json';
 
 export class FortranHoverProvider {
