+++ src/lsp/client.ts
@@ -1,11 +1,15 @@
 'use strict';
 
+import { spawnSync } from 'child_process';
 import * as os from 'os';
 import * as path from 'path';
+
 import * as vscode from 'vscode';
-import { spawnSync } from 'child_process';
 import { commands, window, workspace, TextDocument } from 'vscode';
 import { LanguageClient, LanguageClientOptions, ServerOptions } from 'vscode-languageclient/node';
+
+import { RestartLS } from '../commands/commands';
+import { Logger } from '../services/logging';
 import {
   EXTENSION_ID,
   FortranDocumentSelector,
@@ -15,8 +19,6 @@ import {
   pipInstall,
   resolveVariables,
 } from '../util/tools';
-import { Logger } from '../services/logging';
-import { RestartLS } from '../commands/commands';
 
 // The clients are non member variables of the class because they need to be
 // shared for command registration. The command operates on the client and not
