+++ src/util/tools.ts
@@ -1,8 +1,10 @@
-import * as os from 'os';
-import * as path from 'path';
-import * as vscode from 'vscode';
 import * as assert from 'assert';
 import * as cp from 'child_process';
+import * as os from 'os';
+import * as path from 'path';
+
+import * as vscode from 'vscode';
+
 import { isString, isArrayOfString } from './helper';
 
 export const LS_NAME = 'fortls';
