+++ test/unitTest/client.test.ts
@@ -1,6 +1,8 @@
-import * as vscode from 'vscode';
-import * as path from 'path';
 import { strictEqual } from 'assert';
+import * as path from 'path';
+
+import * as vscode from 'vscode';
+
 import { FortlsClient } from '../../src/lsp/client';
 import { Logger, LogLevel } from '../../src/services/logging';
 import { EXTENSION_ID } from '../../src/util/tools';
