+++ test/unitTest/formatting-provider.test.ts
@@ -1,7 +1,9 @@
 import { strictEqual } from 'assert';
-import { FortranFormattingProvider } from '../../src/format/formatting-provider';
-import * as vscode from 'vscode';
 import * as path from 'path';
+
+import * as vscode from 'vscode';
+
+import { FortranFormattingProvider } from '../../src/format/provider';
 import { Logger, LogLevel } from '../../src/services/logging';
 
 const logger = new Logger(
