+++ test/unitTest/logging.test.ts
@@ -1,4 +1,5 @@
 import { strictEqual } from 'assert';
+
 import { Logger, LogLevel, getConfigLogLevel } from '../../src/services/logging';
 
 suite('Logging', () => {
